Split up CEN files better
